feat: extract shared packages (types + game-logic)
Create @mountain-race/types and @mountain-race/game-logic workspace
packages to share type contracts and simulation logic between the web
client and API server.

- packages/types: Character, GameEvent, GameState, HiddenEffect, Room,
  WebSocket message types (ClientMessage / ServerMessage)
- packages/game-logic: EventSystem, DialogueSystem, balance constants,
  dialogue and event message data
- apps/web: original files become thin re-exports from shared packages,
  all existing import paths remain unchanged
- apps/api: added workspace dependencies for upcoming server simulation
- pnpm check passes (lint + format + typecheck + build)

Made-with: Cursor
